add MAP estimation for lgssm and test
