+++ ssm_jax/distributions.py
@@ -288,7 +288,7 @@ def niw_posterior_update(niw_prior, sufficient_stats):
     
     return NormalInverseWishart(loc=loc_pos, 
                                 mean_concentration=precision_pos, 
-                                df=df_pos, 
+                                df=int(df_pos), 
                                 scale=scale_pos)
 
 
@@ -315,7 +315,7 @@ def mniw_posterior_update(mniw_prior, sufficient_stats):
     
     return MatrixNormalInverseWishart(loc=M_pos, 
                                       col_precision=V_pos, 
-                                      df=nu_pos, 
+                                      df=int(nu_pos), 
                                       scale=Psi_pos)
     
 
@@ -335,7 +335,7 @@ def iw_posterior_update(iw_prior, sufficient_stats):
     df_pos = df_pri + N
     scale_pos = scale_pri + SxxT
     
-    return InverseWishart(df=df_pos, 
+    return InverseWishart(df=int(df_pos), 
                           scale=scale_pos)
