+++ ssm_jax/linear_gaussian_ssm/demos/demos_test.py
@@ -20,4 +20,5 @@ def test_kf_linreg_demo():
 
 
 def test_lgssm_learning_demo():
-    lgssm_learning.main(test_mode=True)
+    lgssm_learning.main(test_mode=True, method='MLE')
+    lgssm_learning.main(test_mode=True, method='MAP')
